+++ components/net/mbedtls/port_ln/include/mbedtls_config.h
@@ -3122,7 +3122,7 @@
  * Uncomment to set the maximum plaintext size of the incoming I/O buffer
  * independently of the outgoing I/O buffer.
  */
-#define MBEDTLS_SSL_IN_CONTENT_LEN              (1024 * 4)
+// #define MBEDTLS_SSL_IN_CONTENT_LEN              (1024 * 4)
 
 /** \def MBEDTLS_SSL_OUT_CONTENT_LEN
  *