+++ components/wifi/wifi_lib_export/wifi.h
@@ -214,11 +214,15 @@ int     wifi_get_channel(uint8_t *channel);
 int     wifi_set_11n_enable(uint8_t ht_enable);
 int     wifi_set_allow_cpu_sleep_flag(uint8_t allow);
 
-int     wifi_sta_set_dtim_period(int period);
+int     wifi_sta_set_listen_interval(int interval);
 int     wifi_sta_get_rssi(int8_t *rssi);
+
+int     wifi_softap_set_dtim_period(int period);
 
 int     wifi_softap_deauth_all(void);
 int     wifi_softap_deauth(uint8_t * mac_addr);
+int     wifi_softap_set_max_supp_sta_num(uint8_t num);
+int     wifi_softap_get_max_supp_sta_num(uint8_t *num);
 
 int     wifi_sta_reg_callback(sta_cb_t *sta_cb);
 int     wifi_softap_reg_callback(ap_cb_t *ap_cb);
@@ -238,8 +242,14 @@ int     wifi_recv_ethernet_pkt_callback_reg(wifi_recv_ethernet_pkt_cb_t recv_cb)
 int     wifi_send_ethernet_pkt(uint8_t *data, int len, uint16_t retry_max, uint8_t retry_timeout);
 int     wifi_private_command(char *pvtcmd);
 
-void    wifi_rf_preprocess(void);
-void    wifi_rf_image_cal(void);
+/**
+ * wifi_rf_calibration
+ * 
+ * Merge function wifi_rf_preprocess() and function wifi_rf_image_cal(),
+ * and finally close RF!
+ * RF will be turned on after mac phy initialization is complete.
+*/
+void    wifi_rf_calibration(void);
 void    wifi_temp_cal_init(uint16_t adc_ch0_val,int8_t cap_comp);
 void    wifi_do_temp_cal_period(uint16_t adc_ch0_val);